Creates editing.integration.tests
